requirements.txt
# LangChain 및 Claude Code 연동을 위한 의존성
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-anthropic==0.1.23
anthropic==0.34.2
openai==1.35.0

# 웹 프레임워크
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0

# Streamlit (메인 웹 인터페이스)
streamlit==1.32.0

# HTTP 클라이언트
httpx==0.27.0
requests==2.32.3

# 환경 변수 관리
python-dotenv==1.0.1

# 로깅
loguru==0.7.2

# 데이터베이스 (Python 3.11 호환)
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# 보안 및 인증
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# 테스트
pytest==8.3.3
pytest-asyncio==0.23.8

# 개발 도구
black==24.8.0
flake8==7.1.1
mypy==1.11.1
